[ET][Portable] Introduce math_util
Differential Revision: [D53624737](https://our.internmc.facebook.com/intern/diff/D53624737/)

[ghstack-poisoned]
