+++ kernels/portable/cpu/op_floor_divide.cpp
@@ -7,6 +7,7 @@
  */
 
 #include <executorch/kernels/portable/cpu/util/broadcast_util.h>
+#include <executorch/kernels/portable/cpu/util/math_util.h>
 #include <executorch/runtime/kernel/kernel_includes.h>
 #include <executorch/runtime/platform/assert.h>
 #include <cmath>
@@ -18,45 +19,6 @@ namespace native {
 
 using Tensor = exec_aten::Tensor;
 using ScalarType = exec_aten::ScalarType;
-
-namespace {
-
-/**
- * Python's __floordiv__ operator is more complicated than just floor(a / b).
- * It aims to maintain the property: a == (a // b) * b + remainder(a, b)
- * which can otherwise fail due to rounding errors in the remainder.
- * So, instead it is calculated as: a // b = (a - remainder(a, b)) / b
- * With some additional fix-ups added to the result.
- */
-template <
-    typename INT_T,
-    typename std::enable_if<std::is_integral<INT_T>::value, bool>::type = true>
-INT_T floor_divide(INT_T a, INT_T b) {
-  const auto quot = a / b;
-  if (std::signbit(a) == std::signbit(b)) {
-    return quot;
-  }
-  const auto rem = a % b;
-  return rem ? quot - 1 : quot;
-}
-
-template <
-    typename FLOAT_T,
-    typename std::enable_if<std::is_floating_point<FLOAT_T>::value, bool>::
-        type = true>
-FLOAT_T floor_divide(FLOAT_T a, FLOAT_T b) {
-  if (b == 0) {
-    return std::signbit(a) ? -INFINITY : INFINITY;
-  }
-  const auto mod = std::fmod(a, b);
-  auto div = (a - mod) / b;
-  if ((mod != 0) && std::signbit(b) != std::signbit(mod)) {
-    return div - 1;
-  }
-  return div;
-}
-
-} // namespace
 
 Tensor& floor_divide_out(
     RuntimeContext& ctx,
@@ -103,8 +65,8 @@ Tensor& floor_divide_out(
                                     static_cast<CTYPE_IN>(val_a);
                                 CTYPE_IN b_casted =
                                     static_cast<CTYPE_IN>(val_b);
-                                CTYPE_IN value =
-                                    floor_divide<CTYPE_IN>(a_casted, b_casted);
+                                CTYPE_IN value = utils::floor_divide<CTYPE_IN>(
+                                    a_casted, b_casted);
 
                                 return static_cast<CTYPE_OUT>(value);
                               },
